+++ CHANGELOG.md
@@ -1,5 +1,21 @@
 # Change log for Microsoft365DSC
 
+# 1.21.512.1
+
+* EXOOfflineAddresBook
+  * Fixed issue in Set-TargetResource where ConfiguredAttributes
+    was passed and resulted in an error.
+* SCDLPComplianceRule
+  * Added several new parameters
+  * Fixed several bugs on extract
+
+# 1.21.505.1
+
+* EXOTransportRule
+  * Adding ExceptIfSCLOver and SCLOver.
+  * Fixes SubjectOrBodyContainsWords parameter not being an array.
+  * Fixes DateTime formating on ExpiryDate and ActivationDate
+
 # 1.21.428.2
 
 * EXOTransportRule